# src/auth/auth.controller.ts
import { Controller, Post, Body, HttpCode, HttpStatus } from '@nestjs/common';
import { AuthService } from './auth.service';
import { UserService } from '../user/user.service';
import { UserCreateDto } from './dto/user-create.dto';
import { UserLoginDto, UserLoginResponseDto } from './dto/user-login.dto';
import { ApiResponse } from '@nestjs/swagger';

@Controller('auth')
export class AuthController {
  constructor(
    private authService: AuthService,
    private userService: UserService,
  ) {}

  @Post('register')
  async register(@Body() createUserDto: UserCreateDto) {
    return this.userService.create(createUserDto);
  }

  @Post('login')
  @HttpCode(HttpStatus.OK)
  @ApiResponse({ status: 200, type: UserLoginResponseDto })
  async login(@Body() loginDto: UserLoginDto) {
    return this.authService.login(loginDto.email, loginDto.password);
  }
}